requirements.txt
streamlit>=1.45.1
python-dotenv>=1.1.0
openrouter>=1.0
openrouter-agent>=0.1.3
openrouter-client>=0.1.0
requests>=2.32.3
pandas>=2.2.3
matplotlib>=3.8.0
seaborn>=0.13.1
numpy>=1.24.0